pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django-cors-headers>=4.7.0",
    "django>=5.1.7",
    "django-ninja>=1.3.0",
    "mysqlclient>=2.2.7",
    "pydantic>=2.10.6",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.0.1",
    "uvicorn>=0.34.0",
    "pillow>=11.1.0",
    "email-validator>=2.2.0",
    "requests>=2.32.3",
]
